+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.1] v0.1.1
+
+* Created `translate` program
+* Updated `summary` program with `--mode` and `--split` flags.
+* Factored out open ai functionality.
+* Updated to top-level await with module exports
+
 ## [0.1.0] v0.1.0
 
 * Added `langchain` package for llm functionality